scripts/research/pyproject.toml
[project]
name = "research"
version = "0.1.0"
description = "LLM-optimized research CLI for web, GitHub repos, and PDFs"
requires-python = ">=3.13"
dependencies = ["click>=8.1", "linkup-sdk", "diskcache"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["research"]
